refactor(podman): grouping cleanup files
