+++ extensions/podman/packages/extension/src/cleanup/podman-cleanup-macos.spec.ts
