+++ extensions/podman/packages/extension/src/cleanup/podman-cleanup-windows.spec.ts
